Fixed input bug + collision bug
Alright, which one of you nimrods decided not to tell me there was an option to have BOTH the og control system AND unity's input system??? And basically all the chunking out i did was for naught? Cause im seconds from throwing hands. Also whoopsie, guess who forgot to put a collider on her player character, eto, bleh. Now that THAT nonsense is done with, I got 2 and a half hours to see about finishing this camera before we hit 7 and any more progress for today is shot. Wish me luck