Implimented FPS movement
Like, it feels like a week of struggle just went down the drain with that revelation. What's the use in continuing when I have what i need already, It's just able to be used in tandem with my regular code now. Sunk cost fallacy going in let's goooooooo
also - 5-3 mofos let's get it